SelfCachingCall.result() - if race condition, sleep random.
Don't sleep for 1 second; sleep for a random number (from 0.5 to 1.4,
with increments of 0.1, inclusive). That way, if there's a true race
condition, we won't *all* be pausing for precisely one second. Each
thread will pause for a slightly different amount of time. This makes an
ongoing race condition less likely.